refactor(apps): Create BaseActivity & clean BaseFragment & BaseFragmentVM
