+++ app/src/main/java/com/franzandel/dicodingexpertsubmission/presentation/activity/DashboardActivity.kt
@@ -1,34 +1,40 @@
 package com.franzandel.dicodingexpertsubmission.presentation.activity
 
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.setupWithNavController
+import com.franzandel.core.presentation.activity.BaseActivity
 import com.franzandel.dicodingexpertsubmission.R
 import com.franzandel.dicodingexpertsubmission.databinding.ActivityDashboardBinding
 import dagger.hilt.android.AndroidEntryPoint
 
+//@AndroidEntryPoint
+//class DashboardActivity : AppCompatActivity() {
+//
+//    private lateinit var binding: ActivityDashboardBinding
+//
+//    override fun onCreate(savedInstanceState: Bundle?) {
+//        super.onCreate(savedInstanceState)
+//        binding = ActivityDashboardBinding.inflate(layoutInflater)
+//        setContentView(binding.root)
+//
+//        setupBottomNavigation()
+//    }
+//
+//    private fun setupBottomNavigation() {
+//        val navHostFragment =
+//            supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
+//        val navController = navHostFragment.navController
+//        binding.navView.setupWithNavController(navController)
+//    }
+//}
+
 @AndroidEntryPoint
-class DashboardActivity : AppCompatActivity() {
+class DashboardActivity : BaseActivity<ActivityDashboardBinding>() {
 
-    private lateinit var binding: ActivityDashboardBinding
+    override fun getViewBinding(): ActivityDashboardBinding =
+        ActivityDashboardBinding.inflate(layoutInflater)
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-//        DaggerHomeComponent.builder()
-//            .context(this)
-//            .appComponent(
-//                EntryPointAccessors.fromApplication(
-//                    applicationContext,
-//                    AppComponent::class.java
-//                )
-//            )
-//            .build()
-//            .inject(this)
-
-        super.onCreate(savedInstanceState)
-        binding = ActivityDashboardBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
+    override fun onActivityCreated() {
         setupBottomNavigation()
     }
