+++ core/build.gradle
@@ -41,6 +41,9 @@ dependencies {
     api 'androidx.appcompat:appcompat:1.2.0'
     api 'com.google.android.material:material:1.3.0'
 
+    api "com.google.dagger:hilt-android:2.35.1"
+    kapt "com.google.dagger:hilt-compiler:2.35.1"
+
     api 'com.squareup.retrofit2:retrofit:2.9.0'
     api 'com.squareup.retrofit2:converter-gson:2.9.0'
