+++ core/src/main/java/com/franzandel/core/presentation/activity/BaseActivity.kt
@@ -1,0 +1,26 @@
+package com.franzandel.core.presentation.activity
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.viewbinding.ViewBinding
+
+/**
+ * Created by Franz Andel on 19/05/21.
+ * Android Engineer
+ */
+
+abstract class BaseActivity<VB : ViewBinding> : AppCompatActivity() {
+
+    protected lateinit var binding: VB
+
+    abstract fun getViewBinding(): VB
+
+    abstract fun onActivityCreated()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = getViewBinding()
+        setContentView(binding.root)
+        onActivityCreated()
+    }
+}
