+++ feature_home/src/main/java/com/franzandel/feature_home/presentation/adapter/DetailAdapter.kt
@@ -2,7 +2,7 @@ package com.franzandel.feature_home.presentation.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import com.franzandel.core.presentation.BaseAdapter
+import com.franzandel.core.presentation.adapter.BaseAdapter
 import com.franzandel.feature_home.databinding.ItemDetailBinding
 import com.franzandel.feature_home.presentation.diffcallback.DetailDiffCallback
 import com.franzandel.feature_home.presentation.vh.DetailVH
